#71793 - Aplicação de patch das alterações da Grid de produtos de orçamento, na tela de cadastro de orçamento.

+++ Glass.Data/Global/DAL/ProdutoDAO.cs
@@ -3121,8 +3121,8 @@ namespace Glass.Data.DAL
             {
                 id = idOrcamento.Value;
                 tipo = ContainerCalculoDTO.TipoContainer.Orcamento;
-                tipoVenda = OrcamentoDAO.Instance.ObterTipoVenda(sessao, idOrcamento.Value).GetValueOrDefault();
-                idParcela = OrcamentoDAO.Instance.ObterIdParcela(sessao, idOrcamento.Value).GetValueOrDefault();
+                tipoVenda = OrcamentoDAO.Instance.ObterTipoVenda(sessao, (int)idOrcamento.Value).GetValueOrDefault();
+                idParcela = OrcamentoDAO.Instance.ObterIdParcela(sessao, (int)idOrcamento.Value).GetValueOrDefault();
             }
 
             #endregion
